Add files via upload
I tried to integrate the menu with the fsm and ISR for the start stop switch but it didn't work. Just uploading this file in case anybody wants to take a look at the failed attempt. The initial menu messages that are supposed to work on the LCD do not work so yea, its fucked. I will update this file in case I some parts later on, but as it is not working I will not be pushing it onto other files just in case.